final score calculations
