+++ src/components/scs/ScSheet/FinalScoreSection.js
@@ -3,7 +3,7 @@ import { injectIntl } from 'react-intl';
 import {
   determineFinalPercentage,
   determineFinalText
-} from './calculationFunc';
+} from './calculations/helperFunctions';
 
 // Material UI
 import { withStyles, Grid, Typography } from '@material-ui/core';
